Move console form to ui/console/form package
