+++ ui/console/form/form.go
@@ -1,4 +1,4 @@
-package ui
+package form
 
 import (
 	"fmt"
