+++ ui/console/form/form_fields.go
@@ -1,4 +1,4 @@
-package ui
+package form
 
 import (
 	"bufio"
