+++ ui/console/form/form_fields_test.go
@@ -1,4 +1,4 @@
-package ui
+package form
 
 import (
 	"testing"
