+++ ui/console/form/form_test.go
@@ -1,4 +1,4 @@
-package ui
+package form
 
 import (
 	"bytes"
